Add slug to form errors
